Wasm memories shouldn't report to the GC until they're attached to a compiled instance https://bugs.webkit.org/show_bug.cgi?id=277711 rdar://133327754
Reviewed by NOBODY (OOPS!).

Right now we inform the GC that we just allocated a huge slab of memory right when we allocate
the JSWebAssemblyMemory. This tends to cause the GC to trigger immediately since Wasm::Memories
are multiples of 64KB even though the Wasm::Memory is probably not going to be collected. Since
Wasm::Memories tend to be allocated right when a Wasm::Module gets compiled this means we're
spending time on a GC that isn't necessarily going to collect much when we could be compiling
the Wasm::Module.

This patch defers informing the GC until the JSWebAssemblyMemory's inevitable instance is
finished. Since a lot of wasm users get the JSArrayBuffer from the JSWebAssemblyMemory
this patch also adds a new constructor that prevents the JSArrayBuffer from reporting it
allocated the Wasm::Memory's backing store since JSWebAssemblyMemory will eventually do so.

* Source/JavaScriptCore/heap/Heap.cpp:
(JSC::Heap::addReference):
* Source/JavaScriptCore/heap/Heap.h:
* Source/JavaScriptCore/runtime/JSArrayBuffer.cpp:
(JSC::JSArrayBuffer::finishCreation):
(JSC::JSArrayBuffer::createForWasmMemory):
* Source/JavaScriptCore/runtime/JSArrayBuffer.h:
* Source/JavaScriptCore/wasm/js/JSWebAssemblyInstance.cpp:
(JSC::JSWebAssemblyInstance::finalizeCreation):
* Source/JavaScriptCore/wasm/js/JSWebAssemblyMemory.cpp:
(JSC::JSWebAssemblyMemory::buffer):
(JSC::JSWebAssemblyMemory::growSuccessCallback):
(JSC::JSWebAssemblyMemory::finishCreation):
(JSC::JSWebAssemblyMemory::didAttachToInstance):
* Source/JavaScriptCore/wasm/js/JSWebAssemblyMemory.h: