Move SimpleRegisterAllocator to a lock count instead of a lock bit. https://bugs.webkit.org/show_bug.cgi?id=296431 rdar://156603761
Reviewed by NOBODY (OOPS!).

In 297809@main I move BBQ's register allocator into a sharable header. But as
Dan pointed out in that review the code was odd because it permitted nested
locking of registers but the first unlock would completely unlock the register
rather than the last. This patch switches to a lock count system instead.