Add a script to auto sort includes https://bugs.webkit.org/show_bug.cgi?id=304246 rdar://166612374
Reviewed by NOBODY (OOPS!).

Right now the style checker will complain if includes in a file are not
sorted alphabetically. This is fine unless you don't know the alphabet...
Add a tool to help automatically sort the headers in a file so I don't
have to sit a my desk attempting to do insertion sort with an O(n)
alphabetical comparision operator.

* Tools/Scripts/webkitpy/style/sort_includes.py: Added.
I put this in a subdirectory rather than at the Tools/Scripts level
since hopefully this script will eventually just be run as part of
style checking with the results presented for validation.